BUUUULLLLLLEEEEETTTTTTSSSSSS

+++ Source/Engine/Renderer/Renderer.cpp
@@ -30,7 +30,7 @@ namespace Twili
 	{
 		SDL_RenderPresent(m_renderer);
 	}
-	void Renderer::setColor(int r, int g, int b, int a)
+	void Renderer::setColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 	{
 		SDL_SetRenderDrawColor(m_renderer, r, g, b, a);
 	}
